refactor(substrate): move the XSD numeric value tower into sparq-substrate (sq-ev41x) [OPUS-4.8]
Behaviour-neutral code move (epic sq-qonbz, Phase 2): the engine's id-level XSD
numeric tower MOVES from `sparq-engine::exec` into the leaf `sparq-substrate`
crate so the reasoners can later share it without depending on the whole engine.

Moved VERBATIM into `sparq_substrate::numeric` (made `pub` + `#[inline]`):
- the `Num` / `Dec` / `ArithOp` / `RoundMode` types and all their methods
  (rank/f64/to_dec/binop/neg/abs/ceil/floor/round/datatype/lexical/
  canonical_lexical/is_nan/is_zero; Dec parse/align/checked_*/cmp/round_to_int/
  lexical),
- the EXACT XSD lexical parsers `parse_xsd_f32` / `parse_xsd_f64`
  (XSD INF/-INF/NaN + scientific forms) and `Dec::parse_lexical` (scale-
  preserving) — this REPLACES the #1290 scaffold placeholder, whose simplified
  `as_numeric` used `Dec::parse`/`str::parse` and mishandled the XSD specials,
- `as_numeric` (= the engine's old `Num::of_literal`), `num_compare`,
  `apply_f64`, `fmt_xsd_double`, `split_decimal`, `sig_digits`.

`sparq-engine` turns the substrate's default-OFF `numeric` feature ON for itself
and `use`s these as free functions; the lean core/wasm build never pulls the
crate by default. The single engine-private piece that did NOT move is
`num_canonical_term` (it builds the engine-private `Value`); it stays a thin
free helper over the shared `Num::canonical_lexical` + `Num::datatype`.

Zero-overhead contract held: no `Box<dyn>`/`&dyn`/vtable in the moved modules
(grep + clippy); every item is `#[inline]` and monomorphic, so with the
workspace `lto = "fat"` profile cross-crate inlining preserves the engine's
FILTER/BIND/ORDER BY codegen.

Perf- and behaviour-neutral (verified):
- W3C SPARQL conformance UNCHANGED: 1225 pass / 0 fail / 4 documented divergence
  = the 1229 ratchet floor; the 4 divergences are the pre-existing decimal
  lexical-form ones.
- 216 engine lib tests pass (numeric / arithmetic / ORDER BY).
- store_bytes_per_triple / dict_bytes_per_term untouched (no storage/dict
  change); wasm bundle +0.048% same-arch (a deterministic LTO symbol-layout
  shift from the extra crate boundary, not new code), well within the +2% band.

Scope: `compare_values` itself stays engine-side — it routes through the
engine-private `Value` / `Term` / datetime `Timeline` model, far outside the
substrate's id-level numeric charter; moving it would haul the engine's whole
value layer into the leaf crate. Its numeric tier already delegates to the
shared `num_compare`. The full `compare_values` move + the join kernels are
deferred to later epic beads. `publish = false` retained (still engine-only
internal; SKILL + micro-bench land with the join-kernel/reasoner-adoption phase).

🤖 SPARQ agent
